# src/interfaces/http/openApi.json
{ 
  "openapi": "3.0.0",
  "info": {
    "title": "Tennis Trainer API Endpoints",
    "version": "v1"
  },
  "servers": [
    {
      "description": "Dev server",
      "url": "/dev/api"
    },
    {
      "description": "Local server",
      "url": "/api"
    },
    {
      "summary": "Auth Token",
      "description": "Auth Server",
      "url": "https://t6q9vl4ui9.execute-api.us-east-1.amazonaws.com/dev"
    }
  ],
  "paths": {
...
    "/clips": {
      "get": {
        "summary": "List all clips per videoId",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "ListClips",
        "tags": [ "Clips" ],"parameters": [
          {
            "name": "videoId",
            "in": "query",
            "description": "Id of video to filter",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "List of all clips",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "responses": {
        "201": {
          "description": "User created successfully",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "400": {
          "description": "User not created because of validation error",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/defaultResponses/ValidationError"
              }
            }
          }
        }
      }
    },
    "/clips/{videoId}/clip": {
      "post": {
        "summary": "Create Clip",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "CreateClip",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "Id of selected video",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "requestBody": {
          "description": "Create clip from selected video",
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                 
                },
                "example": {
                  "videoId": 3,
                  "clipName": "Clip Test name",
                  "startTime": "94",
                  "endTime": "102",
                  "shotType": "backhand",
                  "hitSpot": "behind baseline",
                  "shotResult": "winner",
                  "shotDirection": "cross-court",
                  "comments":"",
                  "clipType" : "forGoldStandard",
                  "winner": "no",
                  "errorType":"wide",
                  "move" : "return",
                  "rallyLength":"1-4"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Clip created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "responses": {
        "201": {
          "description": "User created successfully",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "400": {
          "description": "User not created because of validation error",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/defaultResponses/ValidationError"
              }
            }
          }
        }
      }
    },
    "/clips/{id}": {
      "get": {
        "summary": "Get single clip data",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "ShowClip",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Clip id",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Clip data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update data of a clip",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "UpdateClip",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Clip id",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "requestBody": {
          "description": "Updated Value here",
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                 
                },
                "example": {
                  "clipName": "Clip Test name",
                  "startTime": "94",
                  "endTime": "102",
                  "shotType": "backhand",
                  "hitSpot": "behind baseline",
                  "shotResult": "winner",
                  "shotDirection": "cross-court",
                  "comments":"",
                  "clipType" : "forGoldStandard",
                  "winner": "no",
                  "errorType":"wide",
                  "move" : "return",
                  "rallyLength":"1-4"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Clip updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a clip",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "DeleteClip",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Clip id",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Clip Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "responses": {
        "201": {
          "description": "Updated successfully",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "400": {
          "description": "User not created because of validation error",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/defaultResponses/ValidationError"
              }
            }
          }
        }
      }
    },
    "/clips/video/details/filter": {
      "get": {
        "summary": "Filter clips",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "GetFilter",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "videoId",
            "in": "query",
            "description": "Id of video to filter",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "set",
            "in": "query",
            "description": "set number of the match",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "game",
            "in": "query",
            "description": "game number of the set",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "shotType",
            "in": "query",
            "description": "shot type of the player",
            "required": false,
            "type": "string",
            "enum": [
              "forehand",
              "backhand",
              "volley",
              "overhead"
            ]
          },
          {
            "name": "hitSpot",
            "in": "query",
            "description": "hitting spot of the player",
            "required": false,
            "type": "string"
          },
          {
            "name": "shotResult",
            "in": "query",
            "description": "point result of the player",
            "required": false,
            "type": "string"
          },
          {
            "name": "shotDirection",
            "in": "query",
            "description": "direction of the shot",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "data": [
                      {
                        "videoName": "Danville_10172019_17:54",
                        "matchType": "match",
                        "location": "Danville",
                        "ClipModels": [
                          {
                            "id": 3,
                            "videoId": 1,
                            "clipName": "Danville_10172019_17:54-from:00:01:34_to:00:01:42",
                            "set": 3,
                            "game": 2,
                            "startTime": "94",
                            "endTime": "102",
                            "currentSetScore": "5-4",
                            "currentGameScore": "40-15",
                            "shotType": "forehand",
                            "moveDirection": "body shot",
                            "hitSpot": "behind baseline",
                            "shotResult": "winner",
                            "smartPattern": true,
                            "extra": "hit ball early",
                            "errorType": "NA",
                            "spin": "NA",
                            "shotDirection": "NA",
                            "speed": "NA",
                            "comments": null
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "responses": {
        "201": {
          "description": "User created successfully",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "400": {
          "description": "User not created because of validation error",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/defaultResponses/ValidationError"
              }
            }
          }
        }
      }
    },
    "/clips/filter": {
      "get": {
        "summary": "Filter clips",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "ClipFilter",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "videoId",
            "in": "query",
            "description": "Id of video to filter",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "shotType",
            "in": "query",
            "description": "shot type of the player",
            "required": false,
            "type": "string",
            "enum": [
              "forehand",
              "backhand",
              "volley",
              "overhead"
            ]
          },
          {
            "name": "hitSpot",
            "in": "query",
            "description": "hitting spot of the player",
            "required": false,
            "type": "string"
          },
          {
            "name": "shotResult",
            "in": "query",
            "description": "result of the shot",
            "required": false,
            "type": "string"
          },
          {
            "name": "shotDirection",
            "in": "query",
            "description": "direction of the shot",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "data": [
                      {
                        "videoName": "Danville_10172019_17:54",
                        "matchType": "match",
                        "location": "Danville",
                        "ClipModels": [
                          {
                            "id": 3,
                            "videoId": 1,
                            "clipName": "Danville_10172019_17:54-from:00:01:34_to:00:01:42",
                            "shotType": "forehand",
                            "moveDirection": "body shot",
                            "hitSpot": "behind baseline",
                            "shotResult": "winner",
                            "errorType": "",
                            "shotDirection": "NA",
                            "comments": null
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "responses": {
        "201": {
          "description": "User created successfully",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "400": {
          "description": "User not created because of validation error",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/defaultResponses/ValidationError"
              }
            }
          }
        }
      }
    },
    "/clips/countfilter": {
      "get": {
        "summary": "Filter clips and count value occurences and percentage",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "ListFilteredClips",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "videoId",
            "in": "query",
            "description": "Id of video to filter",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "set",
            "in": "query",
            "description": "set number of the match",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "responses": {
        "201": {
          "description": "User created successfully",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "400": {
          "description": "User not created because of validation error",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/defaultResponses/ValidationError"
              }
            }
          }
        }
      }
    },
    "/clips/{id}/detect-persons": {
      "get": {
        "summary": "Detect persons from clip",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "DetectPersons",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of selected clip",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Return array of persons image",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/clips/{id}/detectedPerson": {
      "post": {
        "summary": "Save detected person images from AI",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "CreateDetectedPerson",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of selected clip",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "requestBody": {
          "description": "Save multiple detected persons from AI",
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "personImages": {
                    "type": "array",
                    "required": true
                  }
                },
                "example": {
                  "personImages": [
                    {
                      "clipId": 1,
                      "image": "image binary base64",
                      "centroid": "(0.3434, 0.4324)"
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Return array of persons image",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List of detected person images from AI",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "ListDetectedPerson",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of selected clip",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Return array of persons image",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/clips/{clipId}/detectedPerson/{id}/setKeypoints": {
      "post": {
        "summary": "Select detected person for keypoints and mp4 generation.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "SetDetectedPersonKeypoints",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of selected person",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "clipId",
            "in": "path",
            "description": "Id of selected clip",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "requestBody": {
          "description": "Set gold scenario and user",
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "scenarioId": {
                    "type": "integer",
                    "required": true
                  },
                  "userId": {
                    "type": "integer",
                    "required": true
                  }
                },
                "example": {
                  "scenarioId": 1,
                  "userId": 2019131640
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/clips/{clipId}/detectedPerson/{id}/generateKeypointsSignedUrl": {
      "post": {
        "summary": "Generate signedURL for selected Persons generated keypoints.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "GenerateKeypointsSignedUrl",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of selected person",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "clipId",
            "in": "path",
            "description": "Id of selected clip",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "requestBody": {
          "description": "Id of personKeypoint",
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "personKeypointId": {
                    "type": "integer",
                    "required": true
                  }
                },
                "example": {
                  "personKeypointId": 4
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/clips/{clipId}/detectedPerson/{id}/generateVideoSignedUrl": {
      "post": {
        "summary": "Generate signedURL for selected Persons generated skeleton mp4.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "GenerateVideoSignedUrl",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of selected person",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "clipId",
            "in": "path",
            "description": "Id of selected clip",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "requestBody": {
          "description": "Id of personKeypoint",
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "personKeypointId": {
                    "type": "integer",
                    "required": true
                  }
                },
                "example": {
                  "personKeypointId": 4
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/clips/{clipId}/detectedPerson/{id}/generateScore": {
      "post": {
        "summary": "Generate scores based on selected model and detectedPerson keypoints.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "GenerateDetectedPersonScore",
        "tags": [ "Clips" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of selected person",
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "clipId",
            "in": "path",
            "description": "Id of selected clip",
            "type": "integer",
            "format": "int64"
          }
        ],
        "requestBody": {
          "description": "Selected model Id",
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "scenarioId": {
                    "type": "integer"
                  },
                  "userId": {
                    "type": "integer"
                  }
                },
                "example": {
                  "scenarioId": 1,
                  "userId" : 1
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "description": "idToken from login"
      }
    },
    "schemas": {
      "User": {
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "Unique Identifier representing a user",
            "x-primary-key": true
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  }
}